Touchstone.jl Documentation

Touchstone.jl Documentation

Touchstone.jl Documentation

Types and Methods

Index

Types and Methods

Touchstone.DataPoint — Type.

Holds scalar frequency and parameter matrix for one data point.

source

Touchstone.Options — Type.

Holds values from option line.

source

Touchstone.TS — Type.

Holds data for Touchstone file.

source

Touchstone.angs — Function.

angs( ts, p1, p2 )

Gets the vector of the angles in radians of parameters with indices p1, p2 from Touchstone data.

source

Touchstone.dBmags — Function.

dBmags( ts, p1, p2 )

Gets the vector of the magnitudes in dB of parameters with indices p1, p2 from Touchstone data.

source

Touchstone.freqs — Method.

freqs( ts )

Gets the frequency vector from Touchstone data.

source

Touchstone.imags — Function.

imags( ts, p1, p2 )

Gets the vector of the imaginary parts of parameters with indices p1, p2 from Touchstone data.

source

Touchstone.mags — Function.

mags( ts, p1, p2 )

Gets the vector of the magnitudes of parameters with indices p1, p2 from Touchstone data.

source

Touchstone.param — Function.

param( ts, p1, p2 )

Gets the vector of parameters with indices p1, p2 from Touchstone data.

source

Touchstone.params — Method.

params( ts )

Gets the vector of parameter matrices from Touchstone data.

source

Touchstone.parse_touchstone_file — Function.

parse_touchstone_file( filename, [ ports ] )

Returns a TS structure for a valid Touchstone data file. Works for ports = 1, 2, 3 or 4.

source

Touchstone.parse_touchstone_stream — Function.

parse_touchstone_stream( stream, [ ports ] )

Returns a TS structure for a valid Touchstone data stream. Works for ports = 1, 2, 3 or 4.

source

Touchstone.parse_touchstone_string — Function.

parse_touchstone_string( in, [ ports ] )

Returns a TS structure for a valid Touchstone data string. Works for ports = 1, 2, 3 or 4.

source

Touchstone.reals — Function.

reals( ts, p1, p2 )

Gets the vector of the real parts of parameters with indices p1, p2 from Touchstone data.

source

Touchstone.write_touchstone_file — Method.

write_touchstone_file( filename, ts )

Writes formated Touchstone data for a TS structure to a file.

source

Touchstone.write_touchstone_stream — Method.

write_touchstone_stream( stream, ts )

Writes formated Touchstone data for a TS structure to a stream.

source

Touchstone.write_touchstone_string — Method.

write_touchstone_string( ts )

Returns a string with formated Touchstone data for a TS structure.

source

Touchstone.default_format — Constant.

The default data format, if not found in option line.

source

Touchstone.default_format_symbol — Constant.

Default data format symbol to use in option line.

source

Touchstone.default_parameter — Constant.

The default parameter format, if not found in option line.

source

Touchstone.default_parameter_symbol — Constant.

Defaoult parameter format symbo to use in option line.

source

Touchstone.default_resistance — Constant.

The default characteristic impedance, if not found in option line.

source

Touchstone.default_resistance_string — Constant.

Default characteristic iimpedance to use in option line.

source

Touchstone.default_unit_value — Constant.

Default unit value to use in option line.

source

Touchstone.default_units — Constant.

The default unit, if not found in option line.

source

Touchstone.formats — Constant.

Symbols for data format for strings found in option line.

source

Touchstone.formatstrings — Constant.

String for data format symbols in option line.

source

Touchstone.parameters — Constant.

Symbols for different parameter formats for strings found in option line

source

Touchstone.parameterstrings — Constant.

Strings for parameter format symbols in option line.

source

Touchstone.units — Constant.

Values for units for strings found in option line.

source

Touchstone.unitstrings — Constant.

String for unit multiplier in option line

source

Base.:== — Method.

Compares data points.

source

Base.:== — Method.

Compares Options.

source

Base.:== — Method.

Compares Touchstone data.

source

Base.:≈ — Method.

Compares data points approximately.

source

Touchstone.comp2da — Method.

comp2da( z )

Returns a pair with the magnitude in dB and angle in degrees of a complex number.

source

Touchstone.comp2ma — Method.

comp2ma( z )

Returns a pair with the linear magnitude and angle in degrees of a complex number.

source

Touchstone.compstring — Method.

compstring( z )

Returns a pair with the real and imaginary components of a complex number.

source

Touchstone.da2comp — Method.

da2comp( m, a )

Converts a pair of values in dB / angle format to a complex number.

source

Touchstone.dastring — Method.

dastring( z )

Returns a string in dB / angle format of a complex number.

source

Touchstone.is_comment_line — Method.

is_commentline( line )

Checks, if a line is a comment line.

source

Touchstone.is_empy_line — Method.

is_empy_line( line )

Checks, if a line is empty.

source

Touchstone.is_format_option — Method.

is_format_option( option )

Checks, if a string is a valid data format in an option line.

source

Touchstone.is_frequency_unit_option — Method.

is_frequency_unit_option( option )

Checks, if a string is a valid unit in an option line.

source

Touchstone.is_keyword_line — Method.

is_keyword_line( line )

Checks, if a line is a valid keyword line.

TODO: Needed for Version 2.0, which is not implemented for now.

source

Touchstone.is_option_line — Method.

is_option_line( line )

Checks, if a line is a option line.

source

Touchstone.is_parameter_option — Method.

is_parameter_option( option )

Checks, if a string is a valid parameter format in an option line.

source

Touchstone.is_resistance_option — Method.

is_resistance_option( option )

Checks, if a string is a valid resistance in an option line.

source

Touchstone.ma2comp — Method.

ma2comp( m, a )

Converts a pair of values in magnitude / angle format to a complex number.

source

Touchstone.mastring — Method.

mastring( z )

Returns a string in magnitude / angle format of a complex number.

source

Touchstone.pairstring — Method.

pairstring( pair )

Returns a string combining the components of a pair.

source

Touchstone.parse_comment_line — Method.

parse_comment_line( line )

Returns the comment of a comment line.

source

Touchstone.parse_format_option — Method.

parse_format_option( option )

Returns the symbol for a valid data format string in an option line.

source

Touchstone.parse_four_port_data — Function.

parse_four_port_data( line, [ options ] )

Returns a data point structure for a valid line of a four-port Touchstone file.

source

Touchstone.parse_frequency_unit_option — Method.

parse_frequency_unit_option( option )

Returns the multiplier for a unit string in an option line.

source

Touchstone.parse_one_port_data — Function.

parse_one_port_data( line, [ options ] )

Returns a data point structure for a valid line of a one-port Touchstone file.

source

Touchstone.parse_option_line — Method.

parse_option_line( line )

Returns the option structure for a valid option line.

source

Touchstone.parse_parameter_option — Method.

parse_parameter_option( option )

Returns the symbol for a valid parameter format string in an option line.

source

Touchstone.parse_resistance_option — Method.

parse_resistance_option( option )

Returns the value of the characteristic impedance for a valid resistance string in an option line.

source

Touchstone.parse_three_port_data — Function.

parse_three_port_data( line, [ options ] )

Returns a data point structure for a valid line of a three-port Touchstone file.

source

Touchstone.parse_two_port_data — Function.

parse_two_port_data( line, [ options ] )

Returns a data point structure for a valid line of a two-port Touchstone file.

source

Touchstone.write_comment_line — Method.

write_comment_line( comment )

Returns a comment line string for a comment.

source

Touchstone.write_format_option — Function.

write_format_option( [ symbol ] )

Returns an option line string for a valid data format symbol.

source

Touchstone.write_four_port_data — Function.

write_four_port_data( data, [ options ] )

Returns a string for a four-port Touchstone formated line for a data point.

source

Touchstone.write_frequency_unit_option — Function.

write_frequency_unit_option( [ value ] )

Returns an option line string for a valid frequency unit multiplier.

source

Touchstone.write_one_port_data — Function.

write_one_port_data( data, [ options ] )

Returns a string for a one-port Touchstone formated line for a data point.

source

Touchstone.write_option_line — Function.

 write_option_line( [ options ] )

Returns the option line string for an option structure.

source

Touchstone.write_parameter_option — Function.

write_parameter_option( [ symbol ] )

Returns an option line string for a valid parameter option symbol.

source

Touchstone.write_resistance_option — Function.

write_resistance_option( [ resistance ] )

Returns an option line string for the characteristic impedance.

source

Touchstone.write_three_port_data — Function.

write_three_port_data( data, [ options ] )

Returns a string for a three-port Touchstone formated line for a data point.

source

Touchstone.write_two_port_data — Function.

write_two_port_data( data, [ options ] )

Returns a string for a two-port Touchstone formated line for a data point.

source

Index

Touchstone.default_format

Touchstone.default_format_symbol

Touchstone.default_parameter

Touchstone.default_parameter_symbol

Touchstone.default_resistance

Touchstone.default_resistance_string

Touchstone.default_unit_value

Touchstone.default_units

Touchstone.formats

Touchstone.formatstrings

Touchstone.parameters

Touchstone.parameterstrings

Touchstone.units

Touchstone.unitstrings

Touchstone.DataPoint

Touchstone.Options

Touchstone.TS

Base.:==

Base.:==

Base.:==

Base.:≈

Touchstone.angs

Touchstone.comp2da

Touchstone.comp2ma

Touchstone.compstring

Touchstone.dBmags

Touchstone.da2comp

Touchstone.dastring

Touchstone.freqs

Touchstone.imags

Touchstone.is_comment_line

Touchstone.is_empy_line

Touchstone.is_format_option

Touchstone.is_frequency_unit_option

Touchstone.is_keyword_line

Touchstone.is_option_line

Touchstone.is_parameter_option

Touchstone.is_resistance_option

Touchstone.ma2comp

Touchstone.mags

Touchstone.mastring

Touchstone.pairstring

Touchstone.param

Touchstone.params

Touchstone.parse_comment_line

Touchstone.parse_format_option

Touchstone.parse_four_port_data

Touchstone.parse_frequency_unit_option

Touchstone.parse_one_port_data

Touchstone.parse_option_line

Touchstone.parse_parameter_option

Touchstone.parse_resistance_option

Touchstone.parse_three_port_data

Touchstone.parse_touchstone_file

Touchstone.parse_touchstone_stream

Touchstone.parse_touchstone_string

Touchstone.parse_two_port_data

Touchstone.reals

Touchstone.write_comment_line

Touchstone.write_format_option

Touchstone.write_four_port_data

Touchstone.write_frequency_unit_option

Touchstone.write_one_port_data

Touchstone.write_option_line

Touchstone.write_parameter_option

Touchstone.write_resistance_option

Touchstone.write_three_port_data

Touchstone.write_touchstone_file

Touchstone.write_touchstone_stream

Touchstone.write_touchstone_string

Touchstone.write_two_port_data