Internals

Internal Documentation

This page lists all the documented internals of the Documenter module and submodules.

Contents

Internal Documentation

Contents

Index

Internal Package Docs

Index

A list of all internal documentation.

Touchstone.ParseConversions

Touchstone.WriteConversions

Touchstone.default_format

Touchstone.default_format_symbol

Touchstone.default_parameter

Touchstone.default_parameter_symbol

Touchstone.default_resistance

Touchstone.default_resistance_string

Touchstone.default_unit_value

Touchstone.default_units

Touchstone.formats

Touchstone.formatstrings

Touchstone.keywordStringSymbols

Touchstone.parameters

Touchstone.parameterstrings

Touchstone.units

Touchstone.unitstrings

Base.:==

Base.:==

Base.:==

Base.:≈

Touchstone.comp2da

Touchstone.comp2ma

Touchstone.compstring

Touchstone.da2comp

Touchstone.dastring

Touchstone.is_comment_line

Touchstone.is_empy_line

Touchstone.is_format_option

Touchstone.is_frequency_unit_option

Touchstone.is_keyword_line

Touchstone.is_option_line

Touchstone.is_parameter_option

Touchstone.is_resistance_option

Touchstone.ma2comp

Touchstone.mastring

Touchstone.pairstring

Touchstone.parse_comment_line

Touchstone.parse_data

Touchstone.parse_data

Touchstone.parse_format_option

Touchstone.parse_frequency_unit_option

Touchstone.parse_keyword_line

Touchstone.parse_option_line

Touchstone.parse_parameter_option

Touchstone.parse_resistance_option

Touchstone.write_comment_line

Touchstone.write_data

Touchstone.write_format_option

Touchstone.write_frequency_unit_option

Touchstone.write_option_line

Touchstone.write_parameter_option

Touchstone.write_resistance_option

Internal Package Docs

Touchstone.ParseConversions — Constant.

Holds conversion functions for parameter format symbols when parsing.

source

Touchstone.WriteConversions — Constant.

Holds conversion functions for parameter format symbols when writing.

source

Touchstone.default_format — Constant.

The default data format, if not found in option line.

source

Touchstone.default_format_symbol — Constant.

Default data format symbol to use in option line.

source

Touchstone.default_parameter — Constant.

The default parameter format, if not found in option line.

source

Touchstone.default_parameter_symbol — Constant.

Defaoult parameter format symbo to use in option line.

source

Touchstone.default_resistance — Constant.

The default characteristic impedance, if not found in option line.

source

Touchstone.default_resistance_string — Constant.

Default characteristic iimpedance to use in option line.

source

Touchstone.default_unit_value — Constant.

Default unit value to use in option line.

source

Touchstone.default_units — Constant.

The default unit, if not found in option line.

source

Touchstone.formats — Constant.

Symbols for data format for strings found in option line.

source

Touchstone.formatstrings — Constant.

String for data format symbols in option line.

source

Touchstone.keywordStringSymbols — Constant.

Symbols for different keywords

source

Touchstone.parameters — Constant.

Symbols for different parameter formats for strings found in option line

source

Touchstone.parameterstrings — Constant.

Strings for parameter format symbols in option line.

source

Touchstone.units — Constant.

Values for units for strings found in option line.

source

Touchstone.unitstrings — Constant.

String for unit multiplier in option line

source

Base.:== — Method.

Compares data points.

source

Base.:== — Method.

Compares Options.

source

Base.:== — Method.

Compares Touchstone data.

source

Base.:≈ — Method.

Compares data points approximately.

source

Touchstone.comp2da — Method.

comp2da( z )

Returns a pair with the magnitude in dB and angle in degrees of a complex number.

source

Touchstone.comp2ma — Method.

comp2ma( z )

Returns a pair with the linear magnitude and angle in degrees of a complex number.

source

Touchstone.compstring — Method.

compstring( z )

Returns a pair with the real and imaginary components of a complex number.

source

Touchstone.da2comp — Method.

da2comp( m, a )

Converts a pair of values in dB / angle format to a complex number.

source

Touchstone.dastring — Method.

dastring( z )

Returns a string in dB / angle format of a complex number.

source

Touchstone.is_comment_line — Method.

is_commentline( line )

Checks, if a line is a comment line.

source

Touchstone.is_empy_line — Method.

is_empy_line( line )

Checks, if a line is empty.

source

Touchstone.is_format_option — Method.

is_format_option( option )

Checks, if a string is a valid data format in an option line.

source

Touchstone.is_frequency_unit_option — Method.

is_frequency_unit_option( option )

Checks, if a string is a valid unit in an option line.

source

Touchstone.is_keyword_line — Method.

is_keyword_line( line )

Checks, if a line is a valid keyword line.

source

Touchstone.is_option_line — Method.

is_option_line( line )

Checks, if a line is a option line.

source

Touchstone.is_parameter_option — Method.

is_parameter_option( option )

Checks, if a string is a valid parameter format in an option line.

source

Touchstone.is_resistance_option — Method.

is_resistance_option( option )

Checks, if a string is a valid resistance in an option line.

source

Touchstone.ma2comp — Method.

ma2comp( m, a )

Converts a pair of values in magnitude / angle format to a complex number.

source

Touchstone.mastring — Method.

mastring( z )

Returns a string in magnitude / angle format of a complex number.

source

Touchstone.pairstring — Method.

pairstring( pair )

Returns a string combining the components of a pair.

source

Touchstone.parse_comment_line — Method.

parse_comment_line( line )

Returns the comment of a comment line.

source

Touchstone.parse_data — Function.

parse_data( line, N, [ options ] )

Returns a data point structure for a valid line of a one-port Touchstone file.

For 3 or more ports, the line should be the concatenation of 3 or more lines from the Touchstone file.

source

Touchstone.parse_data — Function.

parse_data( vals, N, [ options ] )

Returns a data point structure for a vector of numbers from a one-port Touchstone file.

For 3 or more ports, the array should contain exactly 2N² + 1 values.

source

Touchstone.parse_format_option — Method.

parse_format_option( option )

Returns the symbol for a valid data format string in an option line.

source

Touchstone.parse_frequency_unit_option — Method.

parse_frequency_unit_option( option )

Returns the multiplier for a unit string in an option line.

source

Touchstone.parse_keyword_line — Method.

parse_keyword_line( line )

Returns the keyword parameters for a keyword line.

source

Touchstone.parse_option_line — Method.

parse_option_line( line )

Returns the option structure for a valid option line.

source

Touchstone.parse_parameter_option — Method.

parse_parameter_option( option )

Returns the symbol for a valid parameter format string in an option line.

source

Touchstone.parse_resistance_option — Method.

parse_resistance_option( option )

Returns the value of the characteristic impedance for a valid resistance string in an option line.

source

Touchstone.write_comment_line — Method.

write_comment_line( comment )

Returns a comment line string for a comment.

source

Touchstone.write_data — Function.

write_data( data, N [ options ] )

Returns a string for N-port Touchstone formated lines for a data point.

source

Touchstone.write_format_option — Function.

write_format_option( [ symbol ] )

Returns an option line string for a valid data format symbol.

source

Touchstone.write_frequency_unit_option — Function.

write_frequency_unit_option( [ value ] )

Returns an option line string for a valid frequency unit multiplier.

source

Touchstone.write_option_line — Function.

 write_option_line( [ options ] )

Returns the option line string for an option structure.

source

Touchstone.write_parameter_option — Function.

write_parameter_option( [ symbol ] )

Returns an option line string for a valid parameter option symbol.

source

Touchstone.write_resistance_option — Function.

write_resistance_option( [ resistance ] )

Returns an option line string for the characteristic impedance.

source