Public

Public Documentation

Documentation for Touchstone.jl's public interface.

See Internal Documentation for internal package docs covering all submodules.

Contents

Public Documentation

Contents

Index

Public Interface

Index

Touchstone.DataPoint

Touchstone.Options

Touchstone.TS

Touchstone.angs

Touchstone.dBmags

Touchstone.freqs

Touchstone.imags

Touchstone.mags

Touchstone.param

Touchstone.params

Touchstone.parse_touchstone_file

Touchstone.parse_touchstone_stream

Touchstone.parse_touchstone_string

Touchstone.reals

Touchstone.write_touchstone_file

Touchstone.write_touchstone_stream

Touchstone.write_touchstone_string

Public Interface

Touchstone.DataPoint — Type.

Holds scalar frequency and parameter matrix for one data point.

source

Touchstone.Options — Type.

Holds values from option line.

source

Touchstone.TS — Type.

Holds data for Touchstone file.

source

Touchstone.angs — Function.

angs( ts, p1, p2 )

Gets the vector of the angles in radians of parameters with indices p1, p2 from Touchstone data.

source

Touchstone.dBmags — Function.

dBmags( ts, p1, p2 )

Gets the vector of the magnitudes in dB of parameters with indices p1, p2 from Touchstone data.

source

Touchstone.freqs — Method.

freqs( ts )

Gets the frequency vector from Touchstone data.

source

Touchstone.imags — Function.

imags( ts, p1, p2 )

Gets the vector of the imaginary parts of parameters with indices p1, p2 from Touchstone data.

source

Touchstone.mags — Function.

mags( ts, p1, p2 )

Gets the vector of the magnitudes of parameters with indices p1, p2 from Touchstone data.

source

Touchstone.param — Function.

param( ts, p1, p2 )

Gets the vector of parameters with indices p1, p2 from Touchstone data.

source

Touchstone.params — Method.

params( ts )

Gets the vector of parameter matrices from Touchstone data.

source

Touchstone.parse_touchstone_file — Function.

parse_touchstone_file( filename, [ ports ] )

Returns a TS structure for a valid Touchstone data file. Works for ports = 1, 2, 3 or 4.

source

Touchstone.parse_touchstone_stream — Function.

parse_touchstone_stream( stream, [ ports ] )

Returns a TS structure for a valid Touchstone data stream.

source

Touchstone.parse_touchstone_string — Function.

parse_touchstone_string( in, [ ports ] )

Returns a TS structure for a valid Touchstone data string. Works for ports = 1, 2, 3 or 4.

source

Touchstone.reals — Function.

reals( ts, p1, p2 )

Gets the vector of the real parts of parameters with indices p1, p2 from Touchstone data.

source

Touchstone.write_touchstone_file — Method.

write_touchstone_file( filename, ts )

Writes formated Touchstone data for a TS structure to a file.

source

Touchstone.write_touchstone_stream — Method.

write_touchstone_stream( stream, ts )

Writes formated Touchstone data for a TS structure to a stream.

source

Touchstone.write_touchstone_string — Method.

write_touchstone_string( ts )

Returns a string with formated Touchstone data for a TS structure.

source